Update colony-starter command and test script
