+++ packages/colony-starter/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@colony/colony-starter",
   "version": "1.0.0-beta.5",
-  "description": "A global command for creating starter projects",
+  "description": "A command line interface for developers building with Colony",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
