+++ scripts/colony_starter.js
@@ -1,9 +1,10 @@
 #!/usr/bin/env node
 
+// Import prerequisites
+const chalk = require('chalk');
+const cp = require('child_process');
 const fs = require('fs-extra');
 const path = require('path');
-const cp = require('child_process');
-const chalk = require('chalk');
 
 // Exit without error
 const handleExit = () => {
@@ -25,59 +26,78 @@ const handleError = error => {
 process.on('SIGINT', handleExit);
 process.on('uncaughtException', handleError);
 
-// Log start
-console.log();
-console.log(chalk.cyan('  Starting test install...'));
-console.log();
-
-// Get package name from argument
-const packageName = process.argv[2];
-
-// Define package path and json file
-const rootDir = path.join(__dirname, '..');
-const packsDir = path.join(rootDir, 'packs');
-const buildsDir = path.join(rootDir, 'builds');
-const packagesDir = path.join(rootDir, 'packages');
-const packageDir = path.join(packagesDir, packageName);
-const packageJson = path.join(packageDir, 'package.json');
-
-// Check if the package exists
-if (!fs.existsSync(packageJson)) {
-  console.log();
-  console.log(chalk.red('  Unable to locate the requested package!'));
-  console.log();
-  process.exit(1);
-}
-
-// Ensure packs directory exists
-fs.ensureDirSync(packsDir);
-
-// Pack package
-const packName = cp
-  .execSync(`npm pack`, { cwd: packageDir })
-  .toString()
-  .trim();
-
-// Define path to pack file
-const packPath = path.join(packsDir, packName);
-
-// Move packed starter file to packs directory
-fs.rename(path.join(packageDir, packName), packPath);
-
-// Define starter command path
-const starterScriptPath = path.join(packagesDir, 'colony-starter', 'index.js');
-
-// Ensure builds directory exists
-fs.ensureDirSync(buildsDir);
-
-// Execute starter command
-cp.execSync(
-  `node ${starterScriptPath} --specific ${packPath} ${packageName}`,
-  {
-    cwd: buildsDir,
-    stdio: 'inherit',
-  }
+// Define root and script path
+const rootPath = path.join(__dirname, '..');
+const scriptPath = path.join(
+  rootPath,
+  'packages',
+  'colony-starter',
+  'index.js',
 );
+
+// Define starter script arguments
+const scriptArguments = process.argv.slice(2).join(' ');
+
+// Check if build action
+if (process.argv[2] === 'build') {
+
+  // Log start
+  console.log();
+  console.log(chalk.cyan('  Preparing local test build...'));
+  console.log();
+
+  // Define paths
+  const buildPath = path.join(rootPath, 'build');
+  const packagesPath = path.join(rootPath, 'packages');
+  const packagePath = path.join(packagesPath, process.argv[3]);
+  const packageJSON = path.join(packagePath, 'package.json');
+
+  // Check if package exists
+  if (!fs.existsSync(packageJSON)) {
+    console.log();
+    console.log(chalk.red('  Unable to locate the requested package!'));
+    console.log();
+    process.exit(1);
+  }
+
+  // Ensure build path exists
+  fs.ensureDirSync(buildPath);
+
+  // Pack package
+  const packName = cp
+    .execSync(`npm pack`, { cwd: packagePath })
+    .toString()
+    .trim();
+
+  // Pack path
+  const packPath = path.join(buildPath, packName);
+
+  // Move pack to build directory
+  fs.rename(
+    path.join(packagePath, packName),
+    packPath,
+  );
+
+  // Execute starter script
+  cp.execSync(
+    `node ${scriptPath} ${scriptArguments} --specific ${packPath}`,
+    {
+      cwd: buildPath,
+      stdio: 'inherit',
+    }
+  );
+
+} else {
+
+  // Execute starter script
+  cp.execSync(
+    `node ${scriptPath} ${scriptArguments}`,
+    {
+      stdio: 'inherit',
+    }
+  );
+
+}
 
 // Exit script
 handleExit();
